sensor-generator/requirements.txt
requests==2.32.3
pandas==2.2.3
python-dateutil==2.9.0.post0
tqdm==4.66.5
